requirements.txt
Flask>=3.0,<4.0
numpy>=1.26
pandas>=2.0
scikit-learn>=1.7.2,<1.8
joblib>=1.4
kaggle-environments>=1.14
SQLAlchemy>=2.0
pg8000>=1.31
google-cloud-storage>=2.18
google-cloud-tasks>=2.16
google-cloud-secret-manager>=2.20
